// src/types/brand.ts
export type BRAND = {
  logo?: string;
  name?: string;
  visitors?: number;
  revenues?: string;
  sales?: number;
  conversion?: number;
  price?: string; // Price per kWh
  contractLength?: string; // Duration of the contract
  earlyTerminationFee?: string; // Fee for early termination
  renewableOptions?: string; // Percentage of renewable energy options
  specialProvisions?: string; // Specific terms like fixed or variable rates
  customerReviews?: string; // Rating as string
  additionalFees?: string; // Any additional monthly or one-time fees
  paymentOptions?: string; 
};